Implement Firebase Storage abstraction in ActivityDescriptionActivity and ActivityDescriptionActivityUnregister (we can finally get rid of the FirebaseInteraction file now
